fixed package name (uses new dbresearch prefix now)
