+++ src/main/java/at/unisalzburg/dbresearch/apted/distance/APTED.java
@@ -21,15 +21,15 @@
  * SOFTWARE.
  */
 
-package at.unisalzburg.apted.distance;
+package at.unisalzburg.dbresearch.apted.distance;
 
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Stack;
-import at.unisalzburg.apted.node.Node;
-import at.unisalzburg.apted.node.NodeIndexer;
-import at.unisalzburg.apted.costmodel.CostModel;
+import at.unisalzburg.dbresearch.apted.node.Node;
+import at.unisalzburg.dbresearch.apted.node.NodeIndexer;
+import at.unisalzburg.dbresearch.apted.costmodel.CostModel;
 
 /**
  * Implements APTED algorithm [1,2].
@@ -77,14 +77,14 @@ public class APTED<C extends CostModel, D> {
   /**
    * Indexer of the source tree.
    *
-   * @see at.unisalzburg.apted.node.NodeIndexer
+   * @see NodeIndexer
    */
   private NodeIndexer it1;
 
   /**
    * Indexer of the destination tree.
    *
-   * @see at.unisalzburg.apted.node.NodeIndexer
+   * @see NodeIndexer
    */
   private NodeIndexer it2;
