+++ src/main/java/at/unisalzburg/dbresearch/apted/node/Node.java
@@ -21,7 +21,7 @@
  * SOFTWARE.
  */
 
-package at.unisalzburg.apted.node;
+package at.unisalzburg.dbresearch.apted.node;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -87,7 +87,7 @@ public class Node<D> {
   /**
    * Returns a string representation of the tree in bracket notation.
    *
-   * <p>IMPORTANT: Works only for nodes storing {@link at.unisalzburg.apted.node.StringNodeData}
+   * <p>IMPORTANT: Works only for nodes storing {@link StringNodeData}
    * due to using {@link node.StringNodeData#getLabel()}.
    *
    * @return tree in bracket notation.
